+++ src/Miningcore/Persistence/Postgres/Entities/PoolState.cs
@@ -5,7 +5,7 @@ namespace Miningcore.Postgres.Entities
     public class PoolState
     {
         public string PoolId { get; set; }
-        public double HashValue { get; set; }
+        public Decimal HashValue { get; set; }
         public DateTime LastPayout { get; set; }
     }
 }
